requirements.txt
chardet==5.2.0
pillow==10.3.0
reportlab==4.2.0
python-dateutil==2.8.2